+++ crates/mempool/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod communication;
 pub mod mempool;
-pub(crate) mod priority_queue;
+pub(crate) mod transaction_queue;
 // TODO(Mohammad): change to `pub(crate)`.
 pub mod transaction_pool;
